+++ NEWS.md
@@ -1,3 +1,12 @@
+# rsthemes 0.0.3
+
+* Choose theme installation directory based on RStudio version. For 1.3.555+ the
+  themes are now installed into `~/.config/rstudio/themes`. An additional
+  `destdir` argument was added to `install_rsthemes()` for future flexibility.
+  Related to this RStudio desktop issue: 
+  https://github.com/rstudio/rstudio/issues/5674#issuecomment-548484300.
+  (thanks @pat-s)
+
 # rsthemes 0.0.2
 
 * Added automatic light/dark mode switching with `use_theme_auto()` and